pythonService/src/requirements.txt
Flask==3.1.1
kafka_python==2.2.15
langchain_core==0.3.72
langchain_google_genai==2.1.8
langchain_mistralai==0.2.11
langchain_openai==0.3.28
Pillow==11.3.0
protobuf
google-generativeai
pydantic
python-dotenv==1.1.1

